Move the launcher back to SPM and re-add it to the main app target.
