+++ src/cordova/apple/OutlineAppleLib/Package.swift
@@ -12,6 +12,10 @@ let package = Package(
             targets: ["Tun2socks", "OutlineSentryLogger", "OutlineTunnel", "OutlineCatalystApp", "OutlineNotification"]
         ),
         .library(
+            name: "OutlineLauncher",
+            targets: ["OutlineLauncher"]
+        ),
+        .library(
             name: "PacketTunnelProvider",
             targets: ["PacketTunnelProvider"]
         ),
@@ -21,6 +25,14 @@ let package = Package(
         .package(url: "https://github.com/getsentry/sentry-cocoa", from: "7.31.3"),
     ],
     targets: [
+        .target(
+            name: "OutlineLauncher",
+            dependencies: [
+                "CocoaLumberjack",
+                .product(name: "CocoaLumberjackSwift", package: "CocoaLumberjack"),
+                "OutlineCatalystApp",
+            ]
+        ),
         .target(
             name: "OutlineCatalystApp",
             dependencies: [
